Add brace initialization examples for complex and vector classes

+++ src/ch02/vector.h
@@ -8,27 +8,28 @@
 
 class vector {
 public:
-  explicit vector(int size) : my_size{size}, data{new double[size]} {}
+  explicit vector(unsigned size) : my_size{size}, data{new double[size]} {}
   vector(const vector &v) : my_size{v.my_size}, data{new double[my_size]} {
-    for (int i = 0; i < my_size; i++)
+    for (unsigned i = 0; i < my_size; i++)
       data[i] = v.data[i];
   }
   vector(std::initializer_list<double> values)
-      : my_size{static_cast<int>(values.size())}, data{new double[my_size]} {
+      : my_size{static_cast<unsigned>(values.size())},
+        data{new double[my_size]} {
     std::copy(begin(values), end(values), data);
   }
 
-  double &operator[](int index) const { return data[index]; }
+  double &operator[](unsigned index) const { return data[index]; }
   vector &operator=(const vector &src) {
     if (this == &src)
       return *this;
     assert(my_size == src.my_size);
-    for (int i = 0; i < my_size; ++i)
+    for (unsigned i = 0; i < my_size; ++i)
       data[i] = src.data[i];
     return *this;
   }
   vector &operator=(std::initializer_list<double> values) {
-    assert(static_cast<unsigned int>(my_size) == values.size());
+    assert(my_size == values.size());
     std::copy(begin(values), end(values), data);
     return *this;
   }
@@ -39,7 +40,7 @@ public:
   ~vector() { delete[] data; }
 
 private:
-  int my_size;
+  unsigned my_size;
   double *data;
 };
 
@@ -47,7 +48,7 @@ inline double dot(const vector &v, const vector &w) {
   assert(v.my_size == w.my_size);
   double result = 0.0;
 
-  for (int i = 0; i < v.my_size; i++) {
+  for (unsigned i = 0; i < v.my_size; i++) {
     result += v[i] * w[i];
   }
 
@@ -57,7 +58,7 @@ inline double dot(const vector &v, const vector &w) {
 inline std::ostream &operator<<(std::ostream &out, const vector &v) {
   out << '[';
 
-  for (int i = 0; i < v.my_size; i++) {
+  for (unsigned i = 0; i < v.my_size; i++) {
     out << v.data[i] << (i < v.my_size - 1 ? ", " : "");
   }
